refactor: move StatefulRestTemplate class to a visible test helpers package
