+++ operate/common/src/main/java/io/camunda/operate/testhelpers/StatefulRestTemplate.java
@@ -5,10 +5,11 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.operate.util.rest;
+package io.camunda.operate.testhelpers;
 
 import static org.springframework.beans.factory.config.BeanDefinition.SCOPE_PROTOTYPE;
 
+import io.camunda.operate.util.rest.StatefulHttpComponentsClientHttpRequestFactory;
 import java.lang.reflect.Type;
 import java.net.MalformedURLException;
 import java.net.URI;
@@ -26,7 +27,6 @@ import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
 import org.apache.hc.client5.http.protocol.HttpClientContext;
 import org.apache.hc.core5.http.protocol.BasicHttpContext;
 import org.apache.hc.core5.http.protocol.HttpContext;
-import org.springframework.context.annotation.Profile;
 import org.springframework.context.annotation.Scope;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -47,7 +47,6 @@ import org.springframework.web.client.RestTemplate;
  */
 @Component
 @Scope(SCOPE_PROTOTYPE)
-@Profile("dev")
 public class StatefulRestTemplate extends RestTemplate {
 
   private static final String CSRF_TOKEN_HEADER_NAME = "X-CSRF-TOKEN";
