+++ operate/common/src/main/java/io/camunda/operate/testhelpers/StatefultRestTemplateConfiguration.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.operate.util.rest;
+package io.camunda.operate.testhelpers;
 
 import java.util.function.BiFunction;
 import org.springframework.beans.factory.annotation.Value;
