+++ operate/data-generator/src/main/java/io/camunda/operate/data/develop/DevelopDataGenerator.java
@@ -14,8 +14,8 @@ import io.camunda.client.CamundaClient;
 import io.camunda.client.api.worker.JobWorker;
 import io.camunda.operate.data.usertest.UserTestDataGenerator;
 import io.camunda.operate.exceptions.OperateRuntimeException;
+import io.camunda.operate.testhelpers.StatefulRestTemplate;
 import io.camunda.operate.util.ZeebeTestUtil;
-import io.camunda.operate.util.rest.StatefulRestTemplate;
 import io.camunda.webapps.schema.entities.operation.OperationType;
 import java.time.Duration;
 import java.util.ArrayList;
