+++ operate/qa/backup-restore-tests/src/test/java/io/camunda/operate/qa/backup/DataGenerator.java
@@ -15,9 +15,9 @@ import static org.elasticsearch.index.query.QueryBuilders.termQuery;
 import io.camunda.client.CamundaClient;
 import io.camunda.operate.exceptions.OperateRuntimeException;
 import io.camunda.operate.qa.util.ZeebeTestUtil;
+import io.camunda.operate.testhelpers.StatefulRestTemplate;
 import io.camunda.operate.util.RetryOperation;
 import io.camunda.operate.util.ThreadUtil;
-import io.camunda.operate.util.rest.StatefulRestTemplate;
 import io.camunda.operate.webapp.rest.dto.ProcessGroupDto;
 import io.camunda.operate.webapp.rest.dto.SequenceFlowDto;
 import io.camunda.operate.webapp.rest.dto.listview.ListViewResponseDto;
